Simple Tella Switch with message
