+++ mobile/src/main/res/values/colors.xml
@@ -29,7 +29,7 @@
 
     <color name="colorPrimary">@color/wa_black</color>
     <color name="colorPrimaryDark">@color/wa_dark_gray</color>
-    <color name="colorAccent">@color/wa_black</color>
+    <color name="colorAccent">@color/green</color>
     <color name="colorPrimaryInverse">@color/wa_white</color>
 
     <!-- Collect -->
