+++ shared-ui/src/main/res/drawable/rounded_light_purple_background.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
+    <item>
+        <shape>
+            <solid android:color="@color/light_purple" />
+            <corners android:radius="15dp" />
+        </shape>
+    </item>
+</layer-list>
