+++ shared-ui/src/main/res/layout/switch_with_text.xml
@@ -1,64 +1,44 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout
+<androidx.appcompat.widget.LinearLayoutCompat
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:id="@+id/root"
+    android:padding="12dp"
+    android:weightSum="1"
+    android:orientation="horizontal"
     android:layout_width="match_parent"
-    android:background="@color/wa_purple"
     android:layout_height="wrap_content">
 
-    <ImageView
-        android:id="@+id/startImage"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="26dp"
-        android:background="@drawable/tella_dark__logo"
-        android:visibility="gone"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+    <androidx.appcompat.widget.LinearLayoutCompat
+        android:orientation="vertical"
+        android:layout_width="0dp"
+        android:layout_weight=".87"
+        android:layout_height="wrap_content">
 
-    <TextView
-        android:id="@+id/titleTV"
-        style="@style/Tella_Main_White_Text.Bold"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:gravity="start"
-        android:textSize="16sp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toEndOf="@+id/startImage"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_chainStyle="packed" />
+        <TextView
+            android:id="@+id/titleTV"
+            android:gravity="start"
+            style="@style/SettingsTitleText"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:padding="4dp" />
 
-    <ImageView
-        android:id="@+id/endImage"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="26dp"
-        android:background="@drawable/tella_dark__logo"
-        android:visibility="gone"
-        app:layout_constraintStart_toEndOf="@+id/titleTV"
-        app:layout_constraintTop_toTopOf="parent" />
+          <TextView
+              android:id="@+id/explainTV"
+              style="@style/SettingsMessageText"
+              android:layout_width="wrap_content"
+              android:layout_height="wrap_content"
+              android:padding="4dp" />
+
+    </androidx.appcompat.widget.LinearLayoutCompat>
 
     <androidx.appcompat.widget.SwitchCompat
+        style="@style/SwitchWithMessage"
         android:id="@+id/mSwitch"
-        android:layout_width="wrap_content"
+        android:layout_gravity="center_vertical"
+        android:layout_width="0dp"
+        android:layout_weight=".13"
         android:layout_height="wrap_content"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        android:gravity="end"
-        app:layout_constraintVertical_bias="0.5"
-        app:layout_constraintStart_toEndOf="@+id/endImage"/>
+ />
 
-    <TextView
-        android:id="@+id/explainTV"
-        style="@style/Tella_Main_White_Text.Bold"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="21dp"
-        android:textSize="16sp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toStartOf="@+id/mSwitch"
-        app:layout_constraintVertical_chainStyle="packed" />
-</androidx.constraintlayout.widget.ConstraintLayout>
+  </androidx.appcompat.widget.LinearLayoutCompat>
